Rollback get_cache_entry_from_steps
