+++ services/api/tests/routes/test_endpoint.py
@@ -1,12 +1,20 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2022 The HuggingFace Authors.
 
+from http import HTTPStatus
+from unittest.mock import patch
+
+from libapi.exceptions import ResponseNotReadyError
 from libcommon.config import ProcessingGraphConfig
 from libcommon.processing_graph import ProcessingGraph
+from libcommon.queue import Queue
+from libcommon.simple_cache import upsert_response
 from pytest import raises
 
-from api.config import EndpointConfig
-from api.routes.endpoint import EndpointsDefinition
+from api.config import AppConfig, EndpointConfig
+from api.routes.endpoint import EndpointsDefinition, get_cache_entry_from_steps
+
+CACHE_MAX_DAYS = 90
 
 
 def test_endpoints_definition() -> None:
@@ -75,3 +83,91 @@ def test_endpoints_definition() -> None:
         _ = definition["/parquet-and-dataset-info"]
     with raises(KeyError):
         _ = definition["/config-names"]
+
+
+def test_get_cache_entry_from_steps() -> None:
+    dataset = "dataset"
+    revision = "revision"
+    config = "config"
+
+    app_config = AppConfig.from_env()
+    graph_config = ProcessingGraphConfig()
+    processing_graph = ProcessingGraph(graph_config.specification)
+
+    cache_with_error = "config-split-names-from-streaming"
+    cache_without_error = "config-split-names-from-info"
+
+    step_with_error = processing_graph.get_processing_step(cache_with_error)
+    step_without_error = processing_graph.get_processing_step(cache_without_error)
+
+    upsert_response(
+        kind=cache_without_error,
+        dataset=dataset,
+        config=config,
+        content={},
+        http_status=HTTPStatus.OK,
+    )
+
+    upsert_response(
+        kind=cache_with_error,
+        dataset=dataset,
+        config=config,
+        content={},
+        http_status=HTTPStatus.INTERNAL_SERVER_ERROR,
+    )
+
+    # succeeded result is returned
+    result = get_cache_entry_from_steps(
+        processing_steps=[step_without_error, step_with_error],
+        dataset=dataset,
+        config=config,
+        split=None,
+        processing_graph=processing_graph,
+        hf_endpoint=app_config.common.hf_endpoint,
+        cache_max_days=CACHE_MAX_DAYS,
+    )
+    assert result
+    assert result["http_status"] == HTTPStatus.OK
+
+    # succeeded result is returned even if first step failed
+    result = get_cache_entry_from_steps(
+        processing_steps=[step_with_error, step_without_error],
+        dataset=dataset,
+        config=config,
+        split=None,
+        processing_graph=processing_graph,
+        hf_endpoint=app_config.common.hf_endpoint,
+        cache_max_days=CACHE_MAX_DAYS,
+    )
+    assert result
+    assert result["http_status"] == HTTPStatus.OK
+
+    # error result is returned if all steps failed
+    result = get_cache_entry_from_steps(
+        processing_steps=[step_with_error, step_with_error],
+        dataset=dataset,
+        config=config,
+        split=None,
+        processing_graph=processing_graph,
+        hf_endpoint=app_config.common.hf_endpoint,
+        cache_max_days=CACHE_MAX_DAYS,
+    )
+    assert result
+    assert result["http_status"] == HTTPStatus.INTERNAL_SERVER_ERROR
+
+    # pending job throws exception
+    queue = Queue()
+    queue.add_job(job_type="dataset-split-names", dataset=dataset, revision=revision, config=config, difficulty=50)
+    non_existent_step = processing_graph.get_processing_step("dataset-split-names")
+    with patch("libcommon.dataset.get_dataset_git_revision", return_value=revision):
+        # ^ the dataset does not exist on the Hub, we don't want to raise an issue here
+        with raises(ResponseNotReadyError):
+            get_cache_entry_from_steps(
+                processing_steps=[non_existent_step],
+                dataset=dataset,
+                config=None,
+                split=None,
+                processing_graph=processing_graph,
+                hf_endpoint=app_config.common.hf_endpoint,
+                cache_max_days=CACHE_MAX_DAYS,
+            )
