+++ services/search/src/search/routes/search.py
@@ -8,7 +8,7 @@ import re
 from hashlib import sha1
 from http import HTTPStatus
 from pathlib import Path
-from typing import Optional, Tuple
+from typing import List, Optional, Tuple
 
 import duckdb
 import pyarrow as pa
@@ -24,14 +24,19 @@ from libapi.exceptions import (
 from libapi.utils import (
     Endpoint,
     are_valid_parameters,
-    get_cache_entry_from_steps,
     get_json_api_error_response,
     get_json_error_response,
     get_json_ok_response,
     to_rows_list,
+    try_backfill_dataset,
 )
-from libcommon.processing_graph import ProcessingGraph
+from libcommon.processing_graph import ProcessingGraph, ProcessingStep
 from libcommon.prometheus import StepProfiler
+from libcommon.simple_cache import (
+    CACHED_RESPONSE_NOT_FOUND,
+    CacheEntry,
+    get_best_response,
+)
 from libcommon.storage import StrPath, init_dir
 from libcommon.utils import PaginatedResponse
 from libcommon.viewer_utils.features import (
@@ -98,6 +103,45 @@ def full_text_search(index_file_location: str, query: str, offset: int, length: 
     pa_table = query_result.arrow()
     con.close()
     return (num_total_rows, pa_table)
+
+
+# TODO: Copied from api.routes.endpoint, refactor
+
+
+def get_cache_entry_from_steps(
+    processing_steps: List[ProcessingStep],
+    dataset: str,
+    config: Optional[str],
+    split: Optional[str],
+    processing_graph: ProcessingGraph,
+    cache_max_days: int,
+    hf_endpoint: str,
+    hf_token: Optional[str] = None,
+    hf_timeout_seconds: Optional[float] = None,
+) -> CacheEntry:
+    """Gets the cache from the first successful step in the processing steps list.
+    If no successful result is found, it will return the last one even if it's an error,
+    Checks if job is still in progress by each processing step in case of no entry found.
+    Raises:
+        - [`~utils.ResponseNotFoundError`]
+          if no result is found.
+        - [`~utils.ResponseNotReadyError`]
+          if the response is not ready yet.
+    Returns: the cached record
+    """
+    kinds = [processing_step.cache_kind for processing_step in processing_steps]
+    best_response = get_best_response(kinds=kinds, dataset=dataset, config=config, split=split)
+    if "error_code" in best_response.response and best_response.response["error_code"] == CACHED_RESPONSE_NOT_FOUND:
+        try_backfill_dataset(
+            processing_steps=processing_steps,
+            processing_graph=processing_graph,
+            dataset=dataset,
+            hf_endpoint=hf_endpoint,
+            hf_timeout_seconds=hf_timeout_seconds,
+            hf_token=hf_token,
+            cache_max_days=cache_max_days,
+        )
+    return best_response.response
 
 
 def create_response(
